# contracts/openapi/public-v1.yaml
openapi: 3.1.0
info:
  title: Detection Scheduling Workbench API
  version: 1.0.0
  description: Human-approved scheduling API for an electrical-product testing center.
servers:
  - url: /api/v1
security:
  - OidcSessionCookie: []
paths:
...
  /orders:
    get:
      operationId: listOrders
      parameters: [ { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/PageSize' }, { $ref: '#/components/parameters/Query' } ]
      responses: { '200': { $ref: '#/components/responses/OrderPage' } }
    post:
      operationId: createOrder
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/OrderInput' } } } }
      responses: { '201': { $ref: '#/components/responses/OrderResponse' }, '409': { $ref: '#/components/responses/Problem' } }
  /orders/{order_id}:
    parameters: [ { $ref: '#/components/parameters/OrderId' } ]
    get:
      operationId: getOrder
      responses: { '200': { $ref: '#/components/responses/OrderResponse' }, '404': { $ref: '#/components/responses/Problem' } }
    patch:
      operationId: updateOrder
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/IfMatch' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/OrderPatch' } } } }
      responses: { '200': { $ref: '#/components/responses/OrderResponse' }, '409': { $ref: '#/components/responses/Problem' } }
    delete:
      operationId: cancelOrder
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/IfMatch' } ]
      responses: { '204': { description: Cancelled }, '409': { $ref: '#/components/responses/Problem' } }
  /orders/{order_id}/retests:
    post:
      operationId: createRetest
      parameters: [ { $ref: '#/components/parameters/OrderId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/IfMatch' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateRetestRequest' } } } }
      responses: { '201': { $ref: '#/components/responses/OrderResponse' } }
  /orders/{order_id}/pause:
    post:
      operationId: pauseOrder
      parameters: [ { $ref: '#/components/parameters/OrderId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/IfMatch' } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PauseOrderRequest' } } } }
      responses: { '200': { $ref: '#/components/responses/OrderResponse' }, '409': { $ref: '#/components/responses/Problem' } }
  /orders/{order_id}/resume:
    post:
      operationId: resumeOrder
      parameters: [ { $ref: '#/components/parameters/OrderId' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/IfMatch' } ]
      responses: { '200': { $ref: '#/components/responses/OrderResponse' }, '409': { $ref: '#/components/responses/Problem' } }
...
components:
  securitySchemes:
    OidcSessionCookie:
      type: apiKey
      in: cookie
      name: __Host-public_session
      description: OIDC-authenticated BFF session cookie for same-origin browser requests. The BFF sets this cookie after the OIDC login flow with HttpOnly, Secure, and SameSite protections; the React client does not attach a Bearer token and the browser sends the cookie automatically. The BFF validates the OIDC-authenticated session and derives actor identity, center scope, and roles exclusively from verified claims.
  parameters:
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 25 } }
    Query: { name: q, in: query, schema: { type: string } }
    OrderId: { name: order_id, in: path, required: true, schema: { type: string } }
...
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 1 } }
    IfMatch: { name: If-Match, in: header, required: true, schema: { type: string, minLength: 1 } }
  responses:
    Problem: { description: Problem details, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...
    OrderResponse: { description: Order, content: { application/json: { schema: { $ref: '#/components/schemas/Order' } } } }
...
    OrderPage: { description: Order page, content: { application/json: { schema: { $ref: '#/components/schemas/OrderPage' } } } }
...
  schemas:
    Problem: { type: object, required: [type, title, status], properties: { type: { type: string }, title: { type: string }, status: { type: integer }, detail: { type: string }, instance: { type: string } } }
...
    CertificationType: { type: string, enum: [CCC, CVC, international] }
    OrderStatus: { type: string, enum: [pending_schedule, scheduled, in_progress, paused, completed, cancelled] }
    OrderInput: { type: object, required: [sample_name, sample_quantity, certification_type, priority, promised_finish_time, project_ids], properties: { sample_name: { type: string }, sample_quantity: { type: integer, minimum: 1 }, certification_type: { $ref: '#/components/schemas/CertificationType' }, priority: { type: string, enum: [normal, urgent, vip] }, promised_finish_time: { type: string, format: date-time }, project_ids: { type: array, minItems: 1, items: { type: string } } } }
    OrderPatch: { type: object, minProperties: 1, properties: { sample_name: { type: string }, sample_quantity: { type: integer, minimum: 1 }, certification_type: { $ref: '#/components/schemas/CertificationType' }, priority: { type: string, enum: [normal, urgent, vip] }, promised_finish_time: { type: string, format: date-time }, project_ids: { type: array, items: { type: string } } } }
    CreateRetestRequest: { type: object, required: [reason, project_ids], properties: { reason: { type: string, minLength: 1 }, project_ids: { type: array, minItems: 1, items: { type: string } } } }
    PauseOrderRequest: { type: object, required: [reason], properties: { reason: { type: string, minLength: 1 } } }
...
    Order: { allOf: [ { $ref: '#/components/schemas/OrderInput' }, { type: object, required: [id, status, version, created_at], properties: { id: { type: string }, status: { $ref: '#/components/schemas/OrderStatus' }, version: { type: integer }, created_at: { type: string, format: date-time } } } ] }
    OrderPage: { type: object, required: [items, page, page_size, total], properties: { items: { type: array, items: { $ref: '#/components/schemas/Order' } }, page: { type: integer }, page_size: { type: integer }, total: { type: integer } } }
...