# services/api-go/internal/api/generated/model_certification_type.go
/*
 * Detection Scheduling Workbench API
 *
 * Human-approved scheduling API for an electrical-product testing center.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package generated

type CertificationType string

// List of CertificationType
const (
	CCC           CertificationType = "CCC"
	CVC           CertificationType = "CVC"
	INTERNATIONAL CertificationType = "international"
)