# services/api-go/internal/api/generated/model_order.go
/*
 * Detection Scheduling Workbench API
 *
 * Human-approved scheduling API for an electrical-product testing center.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package generated

import (
	"time"
)

type Order struct {
	SampleName string `json:"sample_name"`

	SampleQuantity int32 `json:"sample_quantity"`

	CertificationType CertificationType `json:"certification_type"`

	Priority string `json:"priority"`

	PromisedFinishTime time.Time `json:"promised_finish_time"`

	ProjectIds []string `json:"project_ids"`

	Id string `json:"id"`

	Status OrderStatus `json:"status"`

	Version int32 `json:"version"`

	CreatedAt time.Time `json:"created_at"`
}