# services/api-go/internal/api/generated/model_order_input.go
/*
 * Detection Scheduling Workbench API
 *
 * Human-approved scheduling API for an electrical-product testing center.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package generated

import (
	"time"
)

type OrderInput struct {
	SampleName string `json:"sample_name"`

	SampleQuantity int32 `json:"sample_quantity"`

	CertificationType CertificationType `json:"certification_type"`

	Priority string `json:"priority"`

	PromisedFinishTime time.Time `json:"promised_finish_time"`

	ProjectIds []string `json:"project_ids"`
}