# services/api-go/internal/api/generated/model_order_patch.go
/*
 * Detection Scheduling Workbench API
 *
 * Human-approved scheduling API for an electrical-product testing center.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package generated

import (
	"time"
)

type OrderPatch struct {
	SampleName string `json:"sample_name,omitempty"`

	SampleQuantity int32 `json:"sample_quantity,omitempty"`

	CertificationType CertificationType `json:"certification_type,omitempty"`

	Priority string `json:"priority,omitempty"`

	PromisedFinishTime time.Time `json:"promised_finish_time,omitempty"`

	ProjectIds []string `json:"project_ids,omitempty"`
}