# services/api-go/internal/api/generated/model_order_status.go
/*
 * Detection Scheduling Workbench API
 *
 * Human-approved scheduling API for an electrical-product testing center.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package generated

type OrderStatus string

// List of OrderStatus
const (
	PENDING_SCHEDULE OrderStatus = "pending_schedule"
	SCHEDULED        OrderStatus = "scheduled"
	IN_PROGRESS      OrderStatus = "in_progress"
	PAUSED           OrderStatus = "paused"
	COMPLETED        OrderStatus = "completed"
	CANCELLED        OrderStatus = "cancelled"
)