# services/api-go/internal/api/generated/model_pause_order_request.go
/*
 * Detection Scheduling Workbench API
 *
 * Human-approved scheduling API for an electrical-product testing center.
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package generated

type PauseOrderRequest struct {
	Reason string `json:"reason"`
}